First commit
Halko Networking Server.
The server functionality now:
When a client joins, the server moves it to a StartHub.
In the StartHub the client can create a new room or join an already created room.
After creating a room the server sends the client it's id. 
If a client joins a room, the server sends the joined client it's id, and all other clients the joined clients information.