Bug fix: Room Creation after failing to create a room.
If a client tried to create a room with a viable name after trying to create a room with the a name that was already in use, the room creation would fail. This would happen only in the situation where the name that was used was for example "test" and the viable name started with the same name ("test"). The reason was that the nxtbuflen in the StartHub didn't update after trying to create the room again.